move modern SafeAreaViewWorkaround
