+++ src/components/HistoryEventDetails/layouts/BaseEventDetails.js
@@ -24,7 +24,7 @@ import styled from 'styled-components/native';
 // Components
 import Icon, { type IconName } from 'components/core/Icon';
 import Image from 'components/Image';
-import SafeAreaView from 'components/modern/SafeAreaViewWorkaround';
+import SafeAreaView from 'components/layout/SafeAreaViewWorkaround';
 import SlideModal from 'components/Modals/SlideModal';
 import Text from 'components/core/Text';
