+++ src/components/modern/BottomModal.js
@@ -22,7 +22,7 @@ import * as React from 'react';
 import styled from 'styled-components/native';
 
 // Components
-import SafeAreaView from 'components/modern/SafeAreaViewWorkaround';
+import SafeAreaView from 'components/layout/SafeAreaViewWorkaround';
 import SlideModal from 'components/Modals/SlideModal';
 import Image from 'components/Image';
 import Text from 'components/core/Text';
